fix(precision): 拆 anchor_verification 避免撞 god-file 门槛 (#151)
sentrux_gate.rs:794 的判据是 loc>800 || (functions>25 && loc>400)。单文件版
anchor_verification.rs 是 775 loc / 33 fn，第一个条件没撞线，第二个撞了
(33>25 且 775>400)，把本仓 god_file_count 从基线的 32 推到 33——本仓
run execute 自扫描当时是绿的，只是因为 no_god_files 在这仓的规则里本来就是
false、ratchet 比的又是那份宽松的 .sentrux/baseline.json，自扫描对这条规则
不构成真实门禁。

按 change_risk/change_agenda 现成的目录约定拆开：mod.rs + tests.rs，
mod.rs 里换成 `#[cfg(test)] mod tests;` 声明。functions 计数器按物理行
前缀识别(`is_function_line`，认 "fn "/"pub(crate) fn " 等前缀开头的行)，
so 挪走的是测试模块自己的 18 个 fn，mod.rs 剩 15 个、loc 掉到 400 以内，
两个条件都不再触发；没有为了压数字而在没有第二条关注点的地方硬拆——
状态/计数类型和文件/符号解析器仍然是同一个概念，留在一个 mod.rs 里。

sentrux scan . 验证：拆分前 285 files / god_file_count 33（此分支相对基线
+1 god file 就是 anchor_verification.rs 自己）；拆分后 285 files /
god_file_count 32，回到基线数字。

门禁：cargo build -p code-intel 干净；cargo test -p code-intel 里
anchor_verification 的 17 个测试原样全过(纯物理挪动，逻辑一行没动)。

Refs #151
